feat: adopt FUNSTACK Router 1.4 route handles for fs-routes route objects
Address review feedback on #176:

- Update @funstack/router to ^1.4.0 (dev/peer/fixture/docs deps). The
  router now passes component-typed routes its route definition as a
  `route` prop, so Client Component pages and layouts receive a route
  object from the router itself and the fs-routes `route` prop is
  uniform across Server and Client Components. The Client Component
  fixture page and e2e now cover useRouteParams(route) with the
  router-provided prop.

- Retype FsRouteObject on the router's new RouteHandle (the declared
  type of the `route` prop) instead of PartialRouteDefinition, and
  update FsRouteComponentProps docs for the uniform behavior.

- Move isClientReference into a shared util module
  (src/util/clientReference.ts).

- Docs: present the Server Component staleness as a temporary
  limitation being worked on (#174), and document that Client
  Components receive `route` from FUNSTACK Router v1.4+.
