+++ packages/docs/package.json
@@ -11,7 +11,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@funstack/router": "^1.2.0",
+    "@funstack/router": "^1.4.0",
     "@funstack/static": "workspace:*",
     "@shikijs/rehype": "^4.4.3",
     "@types/node": "catalog:",
