+++ packages/example-fs-routing/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@funstack/router": "^1.2.0",
+    "@funstack/router": "^1.4.0",
     "@funstack/static": "workspace:*",
     "@types/node": "catalog:",
     "react": "catalog:",
