+++ packages/static/e2e/fixture-fs-routing/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "type": "module",
   "devDependencies": {
-    "@funstack/router": "^1.2.0",
+    "@funstack/router": "^1.4.0",
     "@funstack/static": "workspace:*",
     "@types/react": "^19.2.18",
     "@types/react-dom": "^19.2.4",
