+++ packages/static/package.json
@@ -64,7 +64,7 @@
   "author": "uhyo <uhyo@uhy.ooo>",
   "license": "MIT",
   "devDependencies": {
-    "@funstack/router": "^1.2.0",
+    "@funstack/router": "^1.4.0",
     "@playwright/test": "^1.62.1",
     "@types/node": "catalog:",
     "@types/react": "^19.2.18",
@@ -84,7 +84,7 @@
     "srvx": "^0.12.5"
   },
   "peerDependencies": {
-    "@funstack/router": "^1.2.0",
+    "@funstack/router": "^1.4.0",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
     "vite": "^7.0.0 || ^8.0.0"
