+++ packages/static/src/fs-routes/tree.ts
@@ -1,3 +1,4 @@
+import { isClientReference } from "../util/clientReference";
 import type { FsRouteFile, FsRouteModule, FsRouteTreeNode } from "./types";
 
 /**
@@ -92,21 +93,6 @@ function paramName(segment: string): string {
  */
 function isDynamicSegment(segment: string): boolean {
   return segment.startsWith(":");
-}
-
-const CLIENT_REFERENCE = Symbol.for("react.client.reference");
-
-/**
- * Whether a module export is a client reference, meaning the module is marked
- * `"use client"`. React's `registerClientReference` tags every such export
- * with `$$typeof`.
- */
-export function isClientReference(value: unknown): boolean {
-  return (
-    typeof value === "function" &&
-    "$$typeof" in value &&
-    value.$$typeof === CLIENT_REFERENCE
-  );
 }
 
 /**
